+++ install.sh
@@ -1,1 +1,0 @@
-cp package/XMidiCtrl/win_x64/XMidiCtrl.xpl /mnt/x/X-Plane/plugins/XMidiCtrl/win_x64/