+++ src/profile/map/map_in/map_in_pnp.h
@@ -48,7 +48,7 @@ public:
     void set_time_received();
     void set_time_released();
 
-    void read_config(text_logger &in_log, toml::value &in_data, toml::value &in_config) override;
+    void read_config(text_logger &in_log, toml::value &in_data, device &in_device, toml::value &in_config) override;
     bool check(text_logger &in_log) override;
 
     bool execute(midi_message &in_msg, std::string_view in_sl_value) override;