+++ src/utils/utils.h
@@ -15,51 +15,23 @@
 //   If not, see <https://www.gnu.org/licenses/>.
 //---------------------------------------------------------------------------------------------------------------------
 
-#ifndef MAP_IN_H
-#define MAP_IN_H
+#ifndef UTILS_H
+#define UTILS_H
 
 // Standard
-#include <memory>
+#include <string>
 
 // XMidiCtrl
-#include "label.h"
-#include "map.h"
-#include "midi_message.h"
 #include "text_logger.h"
-#include "types.h"
 #include "xplane.h"
 
 namespace xmidictrl {
 
-class device;
-
-class map_in : public map {
+class utils {
 public:
-    explicit map_in(xplane &in_xp);
-    ~map_in() override = default;
-
-    [[nodiscard]] std::string_view sl() const;
-
-    virtual void read_config(text_logger &in_log, toml::value &in_data, device &in_device, toml::value &in_config);
-
-    virtual bool execute(midi_message &in_msg, std::string_view in_sl_value) = 0;
-
-protected:
-    bool check_sublayer(std::string_view in_sl_value);
-    void toggle_dataref(text_logger &in_log, std::string_view in_dataref, std::vector<std::string> &in_values);
-
-    void display_label(text_logger &in_log, float in_value);
-    void display_label(text_logger &in_log, std::string_view in_value);
-
-private:
-    void read_sublayer(text_logger &in_log, toml::value &in_data);
-    void read_label(text_logger &in_log, toml::value &in_data, toml::value &in_config);
-
-    std::string m_sl {};
-
-    std::unique_ptr<label> m_label {};
+    static bool create_preference_folders(text_logger &in_log, xplane &in_xp);
 };
 
-} // Namespace xmiditrl
+} // Namespace xmidictrl
 
-#endif // MAP_IN_H
+#endif // utils_H